+++ models/card_withdraw.go
@@ -16,7 +16,7 @@ type CardWithdraw struct {
 	AmountEbonus int `json:"amount_ebonus" default:"0"`
 	CardDepositId *uuid.UUID `json:"card_deposit_id"`
 	DateCreated time.Time `json:"date_created" gorm:"default:CURRENT_TIMESTAMP"`
-	AmountDiscountCash float32 `json:"amount_discount_cash" default:"0"`
+	AmountDiscountCash float64 `json:"amount_discount_cash" default:"0"`
 
 }
 
@@ -27,7 +27,7 @@ type CardWithdraw struct {
 	AmountEcoin int `json:"amount_ecoin"`
 	AmountEbonus int `json:"amount_ebonus"`
 	CardDepositId *uuid.UUID `json:"card_deposit_id"`
-	AmountDiscountCash float32 `json:"amount_discount_cash"`
+	AmountDiscountCash float64 `json:"amount_discount_cash"`
 }
 
 func (CardWithdraw) TableName() string {
